Aanpassingen solution structuur zodat alle principes in 1 solution terecht komen
