+++ Common/Crosscutting/Logging/Impl/Logger.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Common.Crosscutting.Logging.Impl
+{
+    public class Logger : ILogger
+    {
+        public void LogError(string message)
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
