+++ Common/Entities/Bestelling.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Common.Entities
+{
+    public class Bestelling
+    {
+    }
+}
