+++ Common/Entities/Gebruiker.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Common.Entities
+{
+    public class Gebruiker
+    {
+        public long Id { get; set; }
+
+        public string Naam { get; set; }
+    }
+}
