+++ Open Close Principle/Client.Refactor/OpenClose.Client.Refactor.csproj
@@ -1,0 +1,14 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>netcoreapp2.1</TargetFramework>
+    <AssemblyName>OpenClose.Client.Refactor</AssemblyName>
+    <RootNamespace>OpenClose.Client.Refactor</RootNamespace>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Implementation.Refactor\OpenClose.Implementation.Refactor.csproj" />
+  </ItemGroup>
+
+</Project>
