+++ Open Close Principle/Client/OpenClose.Client.csproj
@@ -1,0 +1,12 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>netcoreapp2.1</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Implementation\OpenClose.Implementation.csproj" />
+  </ItemGroup>
+
+</Project>
