+++ Open Close Principle/Implementation.Refactor/Controller/IGebruikerController.cs
@@ -1,0 +1,12 @@
+﻿using Common.Entities;
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace OpenClose.Implementation.Refactor.Controller
+{
+    public interface IGebruikerController
+    {
+        long AddGebruiker(Gebruiker gebruiker);
+    }
+}
