+++ Open Close Principle/Implementation.Refactor/Controller/Impl/AuthenticatedGebruikerController.cs
@@ -1,0 +1,15 @@
+﻿using Common.Entities;
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace OpenClose.Implementation.Refactor.Controller.Impl
+{
+    public class AuthenticatedGebruikerController : IGebruikerController
+    {
+        public long AddGebruiker(Gebruiker gebruiker)
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
