+++ Open Close Principle/Implementation/Controller/IGebruikerController.cs
@@ -1,0 +1,12 @@
+﻿using Common.Entities;
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace OpenClose.Implementation.Controller
+{
+    public interface IGebruikerController
+    {
+        long AddGebruiker(Gebruiker gebruiker);
+    }
+}
