--- Open Close Principle/SOLID.Demo/Client.Refactor/Client.Refactor.csproj
@@ -1,12 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Implementation.Refactor\Implementation.Refactor.csproj" />
-  </ItemGroup>
-
-</Project>
