--- Open Close Principle/SOLID.Demo/Client/Client.csproj
@@ -1,12 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Implementation\Implementation.csproj" />
-  </ItemGroup>
-
-</Project>
