--- Open Close Principle/SOLID.Demo/Implementation.Refactor/Controller/IGebruikerController.cs
@@ -1,12 +1,0 @@
-﻿using Implementation.Refactor.Entities;
-using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Implementation.Refactor.Controller
-{
-    public interface IGebruikerController
-    {
-        long AddGebruiker(Gebruiker gebruiker);
-    }
-}
