--- Open Close Principle/SOLID.Demo/Implementation.Refactor/Controller/Impl/AuthenticatedGebruikerController.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using Implementation.Refactor.Entities;
-
-namespace Implementation.Refactor.Controller.Impl
-{
-    public class AuthenticatedGebruikerController : IGebruikerController
-    {
-        public long AddGebruiker(Gebruiker gebruiker)
-        {
-            throw new NotImplementedException();
-        }
-    }
-}
