--- Open Close Principle/SOLID.Demo/Implementation.Refactor/Crosscutting/Logging/ILogger.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Implementation.Refactor.Crosscutting.Logging
-{
-    interface ILogger
-    {
-        void LogError(string message);
-    }
-}
