--- Open Close Principle/SOLID.Demo/Implementation.Refactor/Crosscutting/Logging/Impl/Logger.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Implementation.Refactor.Crosscutting.Logging.Impl
-{
-    public class Logger : ILogger
-    {
-        public void LogError(string message)
-        {
-            throw new NotImplementedException();
-        }
-    }
-}
