--- Open Close Principle/SOLID.Demo/Implementation/Controller/IGebruikerController.cs
@@ -1,12 +1,0 @@
-﻿using Implementation.Entities;
-using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Implementation.Controller
-{
-    public interface IGebruikerController
-    {
-        long AddGebruiker(Gebruiker gebruiker);
-    }
-}
