--- Open Close Principle/SOLID.Demo/Implementation/Crosscutting/Logging/ILogger.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Implementation.Crosscutting.Logging
-{
-
-    public interface ILogger
-    {
-        void LogError(string message);
-    }
-}
