--- Open Close Principle/SOLID.Demo/Implementation/Entities/Gebruiker.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Implementation.Entities
-{
-    public class Gebruiker
-    {
-        public long Id { get; set; }
-
-        public string Naam { get; set; }
-    }
-}
